collector/pyproject.toml
[tool.poetry]
name = "collector"
version = "0.1.0"
description = "Data ingestion service for Autonomous Memecoin AI Agent"
authors = ["Artificialesque <oxtuareg@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8"
aio-pika = "^8.0.0"
psycopg2-binary = "^2.9"
snscrape = "*"
vaderSentiment = "*"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
flake8 = "^3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"